+++ src/widgets/dialogs/ColorPickerDialog.hpp
@@ -108,5 +108,7 @@ private:
     void initColorPicker(LayoutCreator<QWidget> &creator);
     void initSpinBoxes(LayoutCreator<QWidget> &creator);
     void initHtmlColor(LayoutCreator<QWidget> &creator);
+
+    void addShortcuts() override;
 };
 }  // namespace chatterino