+++ src/widgets/helper/SearchPopup.cpp
@@ -6,6 +6,7 @@
 #include <QVBoxLayout>
 
 #include "common/Channel.hpp"
+#include "controllers/hotkeys/HotkeyController.hpp"
 #include "messages/Message.hpp"
 #include "messages/search/AuthorPredicate.hpp"
 #include "messages/search/ChannelPredicate.hpp"
@@ -13,7 +14,6 @@
 #include "messages/search/MessageFlagsPredicate.hpp"
 #include "messages/search/RegexPredicate.hpp"
 #include "messages/search/SubstringPredicate.hpp"
-#include "util/Shortcut.hpp"
 #include "widgets/helper/ChannelView.hpp"
 
 namespace chatterino {
@@ -60,11 +60,32 @@ SearchPopup::SearchPopup(QWidget *parent)
 {
     this->initLayout();
     this->resize(400, 600);
+    this->addShortcuts();
+}
 
-    createShortcut(this, "CTRL+F", [this] {
-        this->searchInput_->setFocus();
-        this->searchInput_->selectAll();
-    });
+void SearchPopup::addShortcuts()
+{
+    HotkeyController::HotkeyMap actions{
+        {"search",
+         [this](std::vector<QString>) -> QString {
+             this->searchInput_->setFocus();
+             this->searchInput_->selectAll();
+             return "";
+         }},
+        {"delete",
+         [this](std::vector<QString>) -> QString {
+             this->close();
+             return "";
+         }},
+
+        {"reject", nullptr},
+        {"accept", nullptr},
+        {"openTab", nullptr},
+        {"scrollPage", nullptr},
+    };
+
+    this->shortcuts_ = getApp()->hotkeys->shortcutsForCategory(
+        HotkeyCategory::PopupWindow, actions, this);
 }
 
 void SearchPopup::setChannelFilters(FilterSetPtr filters)